refactor: Extract RichTextboxHelper class to CustomControl
Split the RichTextboxHelper class definition from ActionForm class to create a new custom control. Also added a ReadonlyRichTextbox base class to allow scripts to display formatted text without using WndProc (safety API needed).
